Refactoring the common bits between BinaryFileBasedCollection and FileBasedCollection into GenericFileBasedCollection
